fet:add sorting

+++ src/components/EquipCard.jsx
@@ -1,24 +1,43 @@
 import { FaEdit, FaEye } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const EquipCard = ({ el, idx, equipments, setEquipments}) => {
     const {_id, name, category, price, rating, stock, pTime, imgurl, description, userName, userEmail} = el || {};
 
     
-    const handleDelete=(id)=>{
-        console.log(id)
-        fetch(`https://equipsport-serverside.vercel.app/equipments/${id}`, {
-            method:'DELETE',
-        })
-        .then(res=>res.json())
-        .then(data=>{
-            console.log(data);
-            const newData = equipments.filter((elem)=>elem?._id!==id);
-            setEquipments(newData);
-        })
-
-    }
+    const handleDelete = (id) => {
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#BDEC2F",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!",
+        }).then((result) => {
+          if (result.isConfirmed) {
+            // deletion
+            fetch(`https://equipsport-serverside.vercel.app/equipments/${id}`, {
+              method: "DELETE",
+            })
+              .then((res) => res.json())
+              .then((data) => {
+                console.log(data);
+                const newData = equipments.filter((elem) => elem?._id !== id);
+                setEquipments(newData);
+      
+                Swal.fire({
+                  title: "Deleted!",
+                  text: "Your file has been deleted.",
+                  icon: "success",
+                });
+              })
+          }
+        });
+      };
+      
 
     return (
         <tr className="hover:bg-gray-100">
